Added Ladder Plant (Placeholder)
Added a ladder plant blueprint to be used for level design, as well as the start of adding the proper model and such in. This blueprint has an adjustable height in the editor- this is the "Segment Number" that shows up under "Default" when an instance of the LadderPlant blueprint is in the scene.

To edit the number of segments:
- Adjust the slider for the Segment Number
- Look at the white ring's height to see how high the plant will come to
- Click 'Update Segments' to make the segments appear

The limit is set to 10 because there appears to be a height limit for the player- any higher and you hit an invisible ceiling