Allow BG music to play while paused
Created a new blueprint that has the background music and noise, with both of them being set as 'UI Sound', which allows them to play while the game is paused. A side effect of this is that it also plays while you're in the editor, for some reason?

Also added a few nodes to make the bg music more muffled and toned down while the game is paused